+ fix bug astar algorithm + fix astarWCR + fix TOA of astarWCR

+++ Weather Constrained Routing/src/tools/GraphBuilder.java
@@ -17,7 +17,7 @@ public class GraphBuilder {
         
         for ( int i = 0; i < nodeCount; ++i ) {
             String nodeId = Integer.toString( g.getNodeCount() );
-            Node n = new Node(nodeId, Math.random()*width, Math.random()*height);
+            Node n = new Node(nodeId, (int)(Math.random()*width), (int)(Math.random()*height));
             g.addNode( n );
         }
         
@@ -25,7 +25,7 @@ public class GraphBuilder {
             for (Node u : g.getNodeMap().values()) {
             	if(u == n) continue;
             	if (Node.distanceTo(n,u) < density) {
-            		Edge e = new Edge(n,u,null,Node.distanceTo(n,u));
+            		Edge e = new Edge(n,u,null,(int)(Node.distanceTo(n,u)));
                     if(!g.hasEdge(e))
                     	g.addEdge( e );
                }
